requirements.txt
speedtest-cli==2.1.2
terminaltables==3.1.0
ping3==2.6.1
requests==2.22.0